[SMA-206] Refactoring: Move modals to components package
