+++ solawi-bid-frontend/src/jsMain/kotlin/org/solyton/solawi/bid/module/user/component/modal/create-user-modal.kt
@@ -1,5 +1,4 @@
-package org.solyton.solawi.bid.module.user.modal
-
+package org.solyton.solawi.bid.module.user.component.modal
 
 import androidx.compose.runtime.*
 import org.evoleq.compose.Markup
@@ -19,30 +18,30 @@ import org.evoleq.optics.storage.nextId
 import org.evoleq.optics.storage.put
 import org.jetbrains.compose.web.css.Color
 import org.jetbrains.compose.web.css.color
-import org.jetbrains.compose.web.dom.Div
-import org.jetbrains.compose.web.dom.ElementScope
-import org.jetbrains.compose.web.dom.PasswordInput
-import org.jetbrains.compose.web.dom.Text
+import org.jetbrains.compose.web.dom.*
 import org.solyton.solawi.bid.application.data.device.DeviceType
 import org.solyton.solawi.bid.application.ui.style.form.fieldDesktopStyle
 import org.solyton.solawi.bid.application.ui.style.form.formDesktopStyle
 import org.solyton.solawi.bid.application.ui.style.form.formLabelDesktopStyle
 import org.solyton.solawi.bid.application.ui.style.form.textInputDesktopStyle
 import org.solyton.solawi.bid.module.bid.component.styles.auctionModalStyles
-import org.solyton.solawi.bid.module.user.data.reader.*
+import org.solyton.solawi.bid.module.user.data.reader.errors
+import org.solyton.solawi.bid.module.user.data.reader.inputs
+import org.solyton.solawi.bid.module.user.data.reader.repeatPassword
 import org.solyton.solawi.bid.module.user.service.PasswordCombinationCheck
 import org.solyton.solawi.bid.module.user.service.onPasswordCombinationValid
 import org.w3c.dom.HTMLElement
+import org.solyton.solawi.bid.module.user.data.reader.password as passwordReader
+import org.solyton.solawi.bid.module.user.data.reader.username as usernameReader
 
 @Markup
 @Suppress("FunctionName")
-fun ChangePasswordModal(
+fun CreateUserModal(
     id: Int,
     texts: Source<Lang.Block>,
     modals: Storage<Modals<Int>>,
     device: Source<DeviceType>,
-    storedPassword: String,
-    setUserData: (password: String) -> Unit,
+    setUserData: (username: String, password: String) -> Unit,
     cancel: ()->Unit,
     update: ()->Unit,
 ): @Composable ElementScope<HTMLElement>.()->Unit = Modal(
@@ -59,109 +58,109 @@ fun ChangePasswordModal(
     texts = texts.emit(),
     styles = auctionModalStyles(device),
 ) {
-    var oldPasswordState by remember{ mutableStateOf("") }
-    var newPasswordState by remember { mutableStateOf("") }
-    var newPasswordRepeatState by remember { mutableStateOf("") }
-    var passwordCombinationCheckState by remember { mutableStateOf<PasswordCombinationCheck>(PasswordCombinationCheck.Empty) }
+    var username by remember{ mutableStateOf("") }
+    var password by remember { mutableStateOf("") }
+    var passwordRepeat by remember { mutableStateOf("") }
+    var passwordCombinationCheck by remember { mutableStateOf<PasswordCombinationCheck>(PasswordCombinationCheck.Empty) }
 
     val inputs = texts * inputs
 
     Vertical {
         Div(attrs = { style { formDesktopStyle() } }) {
             Div(attrs = { style { fieldDesktopStyle() } }) {
-                Label((inputs * oldPassword * title).emit(), id = "oldPassword", labelStyle = formLabelDesktopStyle)
-                PasswordInput(oldPasswordState) {
-                    id("oldPassword")
+                Label((inputs * usernameReader * title).emit(), id = "username", labelStyle = formLabelDesktopStyle)
+                TextInput(username) {
+                    id("username")
                     style { textInputDesktopStyle() }
                     onInput {
-                        oldPasswordState = it.value
-                        passwordCombinationCheckState = onPasswordCombinationValid(
-                            value = newPasswordState,
-                            storedPassword,
-                            oldPasswordState,
-                            newPasswordState,
-                            newPasswordRepeatState
+                        username = it.value
+                        passwordCombinationCheck = onPasswordCombinationValid(
+                            value = password,
+                            null,
+                            null,
+                            password,
+                            passwordRepeat
                         ) {
-                                pw : String -> setUserData(pw)
+                                pw : String -> setUserData(username, pw)
                         }
                     }
                 }
             }
 
             Div(attrs = { style { fieldDesktopStyle() } }) {
-                Label((inputs * newPassword * title).emit(), id = "password", labelStyle = formLabelDesktopStyle)
-                PasswordInput(newPasswordState) {
+                Label((inputs * passwordReader * title).emit(), id = "password", labelStyle = formLabelDesktopStyle)
+                PasswordInput(password) {
                     id("password")
                     style { textInputDesktopStyle() }
                     onInput {
-                        newPasswordState = it.value
-                        passwordCombinationCheckState = onPasswordCombinationValid(
-                            value = newPasswordState,
-                            storedPassword,
-                            oldPasswordState,
-                            newPasswordState,
-                            newPasswordRepeatState
-                        ) {
-                                pw : String -> setUserData(pw)
+                        password = it.value
+                        if(password.isNotBlank() && password == passwordRepeat) {
+                            setUserData(username, password)
                         }
+                        passwordCombinationCheck = onPasswordCombinationValid(
+                            value = password,
+                            null,
+                            null,
+                            password,
+                            passwordRepeat
+                        ) {
+                                pw : String -> setUserData(username, pw)
+                        }
+
                     }
                 }
             }
             Div(attrs = { style { fieldDesktopStyle() } }) {
                 Label((inputs * repeatPassword * title).emit(), id = "repeat-password", labelStyle = formLabelDesktopStyle)
-                PasswordInput(newPasswordRepeatState) {
+                PasswordInput(passwordRepeat) {
                     id("repeat-password")
                     style { textInputDesktopStyle() }
                     onInput {
-                        newPasswordRepeatState = it.value
-                        passwordCombinationCheckState = onPasswordCombinationValid(
-                            value = newPasswordState,
-                            storedPassword,
-                            oldPasswordState,
-                            newPasswordState,
-                            newPasswordRepeatState
+                        passwordRepeat = it.value
+                        passwordCombinationCheck = onPasswordCombinationValid(
+                            value = password,
+                            null,
+                            null,
+                            password,
+                            passwordRepeat
                         ) {
-                            pw : String -> setUserData(pw)
+                                pw : String -> setUserData(username, pw)
                         }
                     }
                 }
             }
 
             val  message: String? = messageFrom(
-                passwordCombinationCheckState,
+                passwordCombinationCheck,
                 texts * errors
             )
-
 
             if(message != null) {
                 Div({ style { color(Color.crimson) } }){
                     Text(message)
                 }
             }
+
         }
     }
-
 }
 
-
-
 @Markup
-fun Storage<Modals<Int>>.showChangePasswordModal(
+fun Storage<Modals<Int>>.showCreateUserModal(
     texts: Source<Lang.Block>,
     device: Source<DeviceType>,
-    storedPassword: String,
-    setUserData: (password: String) -> Unit,
+    setUserData: (username: String, password: String) -> Unit,
     cancel: ()->Unit,
     update: ()->Unit,
 ) = with(nextId()) {
     put(this to ModalData(
         ModalType.Dialog,
-        ChangePasswordModal(
+        CreateUserModal(
             this,
             texts,
-            this@showChangePasswordModal,
+            this@showCreateUserModal,
             device,
-            storedPassword,
+
             setUserData,
             cancel,
             update,
